App init/restore
