+++ app/sagas/selectServer.js
@@ -4,7 +4,8 @@ import { AsyncStorage } from 'react-native';
 import { SERVER } from '../actions/actionsTypes';
 import * as actions from '../actions';
 import { connectRequest } from '../actions/connect';
-import { changedServer, serverSuccess, serverFailure, serverRequest, setServer } from '../actions/server';
+import { serverSuccess, serverFailure, serverRequest, setServer } from '../actions/server';
+import { setRoles } from '../actions/roles';
 import RocketChat from '../lib/rocketchat';
 import database from '../lib/realm';
 import * as NavigationService from '../containers/routes/NavigationService';
@@ -19,12 +20,19 @@ const selectServer = function* selectServer({ server }) {
 
 		// yield RocketChat.disconnect();
 
-		yield put(changedServer(server));
 		yield call([AsyncStorage, 'setItem'], 'currentServer', server);
 		const settings = database.objects('settings');
 		yield put(actions.setAllSettings(RocketChat.parseSettings(settings.slice(0, settings.length))));
 		const permissions = database.objects('permissions');
 		yield put(actions.setAllPermissions(RocketChat.parsePermissions(permissions.slice(0, permissions.length))));
+		const emojis = database.objects('customEmojis');
+		yield put(actions.setCustomEmojis(RocketChat.parseEmojis(emojis.slice(0, emojis.length))));
+		const roles = database.objects('roles');
+		yield put(setRoles(roles.reduce((result, role) => {
+			result[role._id] = role.description;
+			return result;
+		}, {})));
+
 		yield put(connectRequest(server));
 	} catch (e) {
 		console.warn('selectServer', e);
